player moves on the server and the movement is sent to the clients with visual results.

+++ Prototype/src/playerobj.h
@@ -9,7 +9,7 @@ namespace Prototype
 	class PlayerObj : public MovableObj
 	{
 	private:
-		int playerId;
+		size_t playerId;
 	public:
 
 		static const float RECTANGLE_SIZE;
@@ -21,11 +21,11 @@ namespace Prototype
 		bool strafingLeft;
 		bool strafingRight;
 
-		PlayerObj(int playerId, const Pos &pos);
+		PlayerObj(size_t playerId, const Pos &pos);
 
 		Pos getPos() const			{ return pos; }
 		void getRectangle(Rectangle &rectangle) const;
-		int getPlayerId() const		{ return playerId; }
+		size_t getPlayerId() const		{ return playerId; }
 	};
 };
